+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # SCSS-Lint Changelog
 
+## master (unreleased)
+
+* Merge `DeclaredName` and `UsageName` into `NameFormat` linter
+
 ## 0.17.1
 
 * Downgrade `colorize` dependency from `0.6.0` back to `0.5.8`